Reorder some modules
