+++ probdiffeq/_probdiffeq/solvers.py
@@ -1,12 +1,7 @@
 """Solvers."""
 
 from probdiffeq import ssm_impl
-from probdiffeq._probdiffeq import (
-    constraints,
-    estimation_strategies,
-    markov_processes,
-    utilities,
-)
+from probdiffeq._probdiffeq import constraints, estimation_strategies, utilities
 from probdiffeq.backend import func, linalg, np, structs, tree
 from probdiffeq.backend.typing import Any, Array, Callable, Generic, TypeVar
 
@@ -18,7 +13,9 @@ N = TypeVar("N", bound=ssm_impl.AbstractTreeNormal)
 Used to type marginals, for example.
 """
 
-T = TypeVar("T", bound=markov_processes.MarkovSequence | ssm_impl.AbstractTreeNormal)
+T = TypeVar(
+    "T", bound=estimation_strategies.MarkovSequence | ssm_impl.AbstractTreeNormal
+)
 """A type-variable to describe posterior distributions."""
