+++ probdiffeq/probdiffeq.py
@@ -5,11 +5,10 @@ See the tutorials for example use cases.
 
 from probdiffeq._probdiffeq.constraints import *
 from probdiffeq._probdiffeq.error_estimators import *
-from probdiffeq._probdiffeq.estimation_strategies import *
-from probdiffeq._probdiffeq.jacobian_handlers import *
+from probdiffeq._probdiffeq.jacobians import *
 from probdiffeq._probdiffeq.loss_functions import *
-from probdiffeq._probdiffeq.markov_processes import *
-from probdiffeq._probdiffeq.priors import *
+from probdiffeq._probdiffeq.markov_strategies import *
 from probdiffeq._probdiffeq.solvers import *
+from probdiffeq._probdiffeq.ssms_and_priors import *
 from probdiffeq._probdiffeq.state_space_models import *
 from probdiffeq._probdiffeq.utilities import *
